COntinued GC 2.0 changes.

+++ src/gaudi2_src/cast_f16_to_i16_gaudi2.cpp
@@ -69,13 +69,13 @@ tpc_lib_api::GlueCodeReturn Castf16toi16Gaudi2::GetGcDefinitions(
     *    Stage II -  Define index space geometry. In this example the index space matches
     *    the dimensions of the output tensor, up to dim 0.
     **************************************************************************************/
-   unsigned int outputSizes[gcapi::MAX_TENSOR_DIM] = {0};
+   uint64_t outputSizes[gcapi::MAX_TENSOR_DIM] = {0};
 
     memcpy(outputSizes, in_defs->inputTensors[0].geometry.maxSizes, sizeof(outputSizes));
 
     // verify that output feature map dimension are correct
     if (memcmp(in_defs->outputTensors[0].geometry.maxSizes, outputSizes,
-               in_defs->outputTensors[0].geometry.dims * sizeof(unsigned) ) != 0)
+               in_defs->outputTensors[0].geometry.dims * sizeof(uint64_t) ) != 0)
     {
         memcpy(in_defs->outputTensors[0].geometry.maxSizes, in_defs->inputTensors[0].geometry.maxSizes, sizeof(outputSizes));
         return tpc_lib_api::GLUE_INCOMPATIBLE_OUTPUT_SIZE;
